insert fixed

+++ SpringMicroservices/universitySE/src/main/java/com/example/universitySE/repositories/TeachingMaterialRepository.java
@@ -15,6 +15,6 @@ public interface TeachingMaterialRepository extends JpaRepository<TeachingMateri
     @Query(value = "INSERT INTO Teaching_material (subject, media) VALUES (?1, ?2)", nativeQuery = true)
     void addTeachingMaterial(int subject, String media);
 
-    List<TeachingMaterial> getTeachingMaterialBySubject(int subject);
+    List<TeachingMaterial> getTeachingMaterialBySubject(long subject);
 
 }
